refactor: use server ioc
