--- src/google/google.config.interface.ts
@@ -1,6 +1,0 @@
-interface IGoogleConfig {
-  readonly clientId: string;
-  readonly clientSecret: string;
-}
-
-export type { IGoogleConfig };
