--- src/google/index.ts
@@ -1,2 +1,0 @@
-export { GoogleModule, GOOGLE_CONFIG } from "./google.module";
-export type { IGoogleConfig } from "./google.config.interface";
